refactor(thread): 将任务管理器重命名为任务执行器并更新引用
- 将 AsyncTaskManager 重命名为 AsyncTaskExcutor
- 更新 ScheduleMonitor 中的所有引用和文档注释
- 删除原 AsyncTaskManager 接口及其实现类 DefaultAsyncTaskManager
- 新增 AsyncTaskExcutor 接口及其实现类 DefaultAsyncTaskExcutor
- 保持原有功能不变，仅进行名称重构和代码迁移

+++ app/app-platform-chat/providers/ai-provider-ai-agent/src/main/java/ai/platform/aiassit/service/ai/agent/service/AiAgentProvider.java
@@ -604,7 +604,7 @@ public class AiAgentProvider implements AiChatService, AgentRuntime {
     }
 
     private Duration resolveTimeout(Integer requestTimeoutMs) {
-        Integer timeoutMs = requestTimeoutMs != null && requestTimeoutMs > 0
+        Long timeoutMs = requestTimeoutMs != null && requestTimeoutMs > 0
                 ? requestTimeoutMs
                 : properties.getTimeoutMs();
         return Duration.ofMillis(timeoutMs != null && timeoutMs > 0 ? timeoutMs : 60000L);
